Combine MetaDataOnly stories into AssetViewPage
The AssetViewPage has smarts to switch between showing an object viewer or just showing the meta data. The MetaDataOnlyView does not. This means that when clicking a a link to a related asset with image in the MetaDataOnlyView, it wouldn't switch to different layout, which may make devs think the MetaDataOnlyView is broken... but it's just a "dumb" view without routing, so this is expected functionality.

To avoid confusion, this:
- Eliminates a separate story for the MetaDataOnly view, and puts it with the AssetViewPage, which has the smarts needed to switch views on clicks.
- Renames MetaDataOnlyPage to MetaDataOnlyView. The thinking here is that "pages" have smarts for switching "views" if needed.