switch from tailwind config to plain css vars
using tailwind config as middle layer was nice for tailwind autocompletion, but adds an extra step for using the CSS variable: after adding the variable, you need to remember to add it to the tailwind config file (and look up which tailwind config prop it relates too).

The tailwind config stuff seems like it might be too complicated to maintain over time (what happens if tailwind changes its config format?). So, instead, we just use raw CSS variables and classes.