app menu refactor
the previous structure of the app menu lets us use navigational links. However, in some cases, we want menu items that take action on the current asset, like DELETE, and these actions require a confirm modal.

This ditches the NavItemFactory in leiu of more declarative submenus, which should make it easier to add @click handlers in the future.