only mount the Embed button if we're in embed mode
This prevents embed `setup` from running until needed, which keeps the api to fetching interstitial info from firing off.

It may be better avoid getting interstial onMount with the button and wait for the click, but this adds an extra api call when clicking add, which could feel slow to the user?