use separate page for browsing collection
Previously, when browsing a collection, we would navigate to `/collections/browseCollection/:collectionId`, and then once we had the corresponding searchId, we'd navigate to `/search/s/:searchId`.

This put two entries in the browser history, and made navigating back problematic as it would redirect back to the search once it got to the browse collections page.

Replace the history with the `collectionId` was a bit better, and matches the experience on the currently website (I think), but the downside is that when a user "browsing" a collection hits refresh, they just get the previous search results, not necessarily the most up-to-date collection.

In this commit, I separated the Seach and broweCollections pages so that we can have somewhat different functionality. On the browseCollections page, the  user will "browse" page will always get fresh search results.