change to fully controlled tabs
activeTab is now passed as prop, rather than some internal state. This avoids the complexity of arriving on page, setting the intial active tab via a prop, and then keeping the track of the active tab in internal state (but updating the url). instead it's clearer to do one way data flow: searchStore.displayView holds the active tab state, and then passes it down to `<Tabs>`.