inversion of control: MapPopup
Mapbox doesn't have a nice way to grab all attached markers that I could find. So the plan is to invert control: Map will give Markers an `addMarker` function, and Markers will give Popups and `addPopup` function.

The Marker component will be responsible for invoking the `addMarker` method, which will pass the info directly to the `Map`... and then the map can hold onto all markers. And once we have a list of all markers, we can then do the clustering stuff.

Likewise Marker passes it's child Popup a `addPopup` method, which the `MapPopup` can invoke, and then the Marker creates the popup, not the `<Popup>` component.

To do this, we start with the inner most component (the Popup) and then work outward to the Map.

One additional note: `addPopup` returns the function to `removePopup` if needed. Much like `watch` returns a function to stop watching.