omg it was arcgis fonts
The layer for the cluster text (marker count) wasn't rendering and prevented the cluster circle layer from appearing as well.

So basically my map was completely blank, unless you zoomed in.

The issue turned out to be something with fonts and ArcGIS. Switching to their layer style with `Arial Bold` font fixed things.