Clear activeAssetId when leaving asset page
Fixes an issue where the user navigates to an asset page, and then goes to a non-asset page. If the `activeAssetId` remains set, editing controls will be available for this asset, even though the user isn't viewing the asset. By clearing the `activeAssetId` when leaving the asset page, we prevent the menu actions for a particular asset from appearing.