Fix App Menu `Edit` doesn't appear (#115)
* Fix App Menu `Edit` doesn't appear

When there's no asset loaded in the assetStore, the Edit menu won't appear in the App Menu since this `assetId` is null.

This removes the check for a truthy `assetId` from `<AppMenu>` and instead checks within the `EditNavSection` so that users still have an option to Add Assets or view all their assets if they `canManageAssets`

* Clear activeAssetId when leaving asset page

Fixes an issue where the user navigates to an asset page, and then goes to a non-asset page. If the `activeAssetId` remains set, editing controls will be available for this asset, even though the user isn't viewing the asset. By clearing the `activeAssetId` when leaving the asset page, we prevent the menu actions for a particular asset from appearing.

---------

Co-authored-by: Elevator Dev (ubuntu) <latistecharch@umn.edu>