fix duplicate slides added
since props.matches will contain all the slides, if we want to avoid rerendering the whole gallery, we will need to add just the changed slides to the `slides` ref.

This computes the difference between `newMatches` and `oldMatches` and then creates slides for those that aren't in the old set.

Note: I had a difficult time getting `newMatches` and `oldMatches` to show different values or even detecting a change.

I first added `deep: true` to detect any change on props.matches.

And then, to get newMatches to be different from oldMatches, we used the spread operator to make a shallow copy.

> Note: when mutating (rather than replacing) an Object or an Array and watch with deep option, the old value will be the same as new value because they reference the same Object/Array. Vue doesn't keep a copy of the pre-mutate value.

see: https://stackoverflow.com/questions/68463713/vue3-new-value-and-old-value-returned-by-deep-watcher-always-same