fix: old search results remain on new search
This commit is focused on resolving an issue where a user may see old results in the search page's gallery (and possibly elsewhere) when searching with a new query, the old results would remain and new slides would be pushed to the end of the slides.

At first, I tried just having the search results slides component detect when the search changed, and then do a reset. However, this seemed to be brittle and I would encounter cases that wouldn't trigger the reset properly.

Additionally, I worried that other components on the search page would face the same issue.

So, this takes the approach of remounting all the search tabs when the search changes.

This:
- adds some hooks to the searchStore, so that we can subscribe to new searches: `onBeforeNewSearch` and `onAfterNewSearch`.
- adds a new ref `isNewSearchReadyForDisplay` to the SearchResultsPage, which will track if we're clear to render the search results tabs
- subscribes to `onBeforeNewSearch` and `onAfterNewSearch`, toggling the `isNewSearchReadyForDisplay` value from false to true once ready.

Finally, there used to be two functions which would trigger a new search: one for when needed to first get the `searchId`, and another when we already had the `searchId` like when we were already on the page and the `searchId` was in the route.

This commit consolidates the two cases into one `search` function (where providing the searchId is an optional parameter). Then removes the old `searchById` action, so we can just call `search` and trigger our before/after handlers in one function, which avoids needing to worry about triggering before/after handlers more than once or not at all.