handle thumbnail touch and scrollbar slide changes
When a user drags the thumbs or the thumbs scrollbar, update the `activeSlideIndex` and update the `mainSwiper` to scroll to new active slide.

This attaches the handlers to `@touchEnd` and `@scrollbarDragEnd` events rather than `@slideChange` to avoid lots of extraneous object viewer loading in the main slideshow, which gets laggy.