fix grid double-tap issue on iOS safari
Resolves: On mobile search results page, it takes 2 taps to go to the result #95

This issue was caused by CSS changing opacity on a button when hovering. iOS tries to do the smart thing and reveal the new button on first "hover" tap, and then follows the link as expected on the second tap.

This removes the opacity change and changes the button to a div, since the element is just stylistic. The user can click anywhere on the card to go to the asset.