fix: search input empty on search page
when a user enters or refreshes a search page, the search input is empty.

This is because searchStore.query isn't set after the initial results.

We want to be careful not to overwrite existing search query if it's there (e.g. user modifies their query while waiting for the initial search to return).

But, if the query is empty, we should use the `searchText`.