making browsing title flash a bit better
when the sort changes and there's a new search, the searchStore's search entry is cleared so we lose the computed title.

Some attempts to persist until the search completes leads to another problem: mislabelled collections when the user moves from one browsing collection to another.

This doesn't solve the problem, but at least reduces the jarring page shifts on sort change in browsing collections, until  a better solution can be found.