fix menu tab-able when closed
When tabbing through a page, a user would open the menu and then not be able to dismiss it. This was because the menu was just offscreen, not truly hidden.

This changes to using vue's Transition component, and then hiding the menu when it's not open.

Also some accessibilty fixes to focus the Menu button when a close is triggered.