Clear search filters on static pages Fixes #157
Instead of clearing the search filters when leaving the search results page, this clears them when entering a static page – which is usually (always?) what we want.

The other approach would have been to always trigger a search when entering a search results page. But this could cause double requests since one request is triggered on search form submit to get ID), and another one is triggered when the search results page is mounted (in the case a user arrives without an id).