break search into two functions
When getting filters to persist, I noticed searches can a search request multiple times. This is because the search input form submission would trigger a search and arriving on the search page would trigger another.

Takes the `search()` function and breaks into two functions: 1) `getSearchId()`, and 2) `search()` which loads the results in the search store.

Then the submit handler in the searchBar can just be responsible for getting the searchId and then redirected to the searchId.