change clicktosearch to use multiselect query
To avoid doing a full page refresh with the `<a>` tag, we switch to a button that looks like a link.

When clicked, the button will setup the searchStore for a new multiselect field query, and fire it off.

Searching involves a couple of api calls: one to get the searchId, and another to get the results.

The query to get the searchId is noticable, and waiting for it to return before redirecting makes the page feel slugging.

Adding a full screen loading component was a bit disruptive too. Instead, we queue up the the search for the next tick, and redirect to `/blank` while we're waiting for the search to complete.

The blank page immediately, and then the loading results shortly thereafter makes the page feel more responsive.