simplify - remove asset from drawer
ts has trouble with `this` and knowing when `this.what.ever` is defined since the value could possibly change.

We have to use an exclamation point after `what` in `this.what!.ever` to assert that `this.what` is defined. We need to do this even AFTER we've done an explicit check to see if it's defined. And then we need to add an `eslint-disable` which is gross.

Assigning the reference to a different variable will let us skip the `!` asserting, and seems like the more better way to solve the issue since it lets us skip some later checks and shortens the code when we optimistically remove the asset.