fix local login bad credentials handling
The axios interceptor is swallowing the error, which causes a user with a bad username/password to get the Login Required modal on the Login page. This is not what we want – we want the login form to show the error message and not the modal.

To fix this, we use the `skipErrorNotifications` custom config to allow the error to skip the interceptor, and then do the error handling ourselves. The error is now an ApiError and not an AxiosError, so we fix this while keeping the return type the same.