Create an IconButton component
In prep for adding an edit permissions icon button on the drawers page, we create an IconButton component. We have a lot of one-off icon buttons each with a slightly different style. And, the icons can be hard to interpret. Having one component should unify the styles. And Wrapping it with a PopOver component, should provide some text hint as to the purpose of the button.

A lot of the logic within is so that we can handle router links, hrefs, and plain old buttons with click actions all with one component.