fix: search results don't update when sort changes
This fixes a regression (introduced in 17fdc62) where the results on the search page don't update after a user changes the sort order. After updating the searchStore's state with the new sort option, we need to get first get the new searchId, and then use the searchId to grab the results.

Previously, we weren't getting a new searchId after the sort changes.