put submit button in footer
BEfore this the `Add to Drawer` button might only be visible if a user scrolled since the content was larger than the modal. This wasn't ideal because it's possible to see all the form fields EXCEPT the submit button.

This moves the submit button into a fixed footer area while allowing the main modal content to scroll.