fix custom header location
previously was within app header, but this made the folwell theme look weird since the yellow top border would appear above the custom header. Additionally, it was rendering on more than just the static and home page.

This change moves it out of appheader entirely and just includes it in a custom slot.

Also does the same thing with the app footer, which fixes a rendering issue with it.