fix: map marker not rendering if no primaryHandlerId
This resolves an issue where only a subset of results with locations where showing on the map search results page.

The issue was the check for primaryHandlerId: if this was null, the loop skipped the trying to render the match.

`primaryHandlerId` is needed for rendering a thumbURL, but not every match will have a thumbnail. The asset may be metadata only as is the case with the some NIBSDA entries for boarding school locations.

There is still an issue with multiple rerenderings, which could be related to performance issues seen elsewhere on the search results page.