fix recursive rerending warning
When using a `watch` in the MapMarker components, vue will warn that maximum recursive updates exceeded.

This warning only happens with a large data set loaded.

In this commit, I've disabled the watch on the lng/lat so that `createOrUpdateMarker` only runs once.

But, a better fix will be to discover what in `createOrUpdateMarker` is causing all these recursive updates.