stop recursive rerendering
There seems to be a conflict between rerendering when the markers object changes and rerendering on a `styledata` change. (I think this happened when we changed to grouping markers on zoom, I changed how markers were rendered, but didn't remove the old rerender triggers).

This changes so that `styledata` event handler handles the initial rendering and any map style changes, then removes the rerender watch on mapRef and markers and the activeMapStyleKey.

Also, since we don't need to handle marker location changes after initial render, I've removed the watcher on lng and lat within the marker and changed the name of `createOrUpdateMarker` to just `addMarker`.