fix: safari error when downloading a file Fixes #291
When the download link is clicked in safari, the browser cancels any in flight requests – maybe as it prepares to change to a new page? (A 499 is observed on the server.)

This causes some api requests to fail after the download link click.

Changing from a link to a button which invokes the download link sidesteps the issue.