fix: api errors in safari
Here's what I think is happening on safari: if the download request is in flight when the next request begin, the download request will be cancelled giving an "APIError"

To work around this we need to allot enough time for the server to respond to the download request. (It doesn't need to finish downloading, just get a 200 from the server).

Most responses return in less than 1s on dev, but occasionally I'll see a 1.8s response so making this ~2s to give a little overhead.