ignore firstObjectId when it's firstFileHandlerId
This is a (hopefully temporary) workaround where sometimes we get a 404 error when viewing an asset.

To trigger:
1. Search for "All Files" on dev
2. Choose "All Files One" asset. Usually has no error. (but sometimes does?)
3. Go back to search results.
4. Choose "All Files Two" asset. 404 will trigger.

You can also do it in reverse.

The issue seems to be we get diferent results from the api, where `firstObjectId: null` sometimes and `firstObjectId: <fileHandlerId>` other times.

This workaround just ignores firstObjectId if it's the same as firstFileHandlerId.

I'm not sure this is the correct fix. activeObjectId determines when we show the detail panel.