fix: download all skips the first file in Safari
This is a workaround for an issue in safari where the first download is (sometimes?) skipped when the user chooses "Download All."

The issue is related to the caching of `getFileDownloadInfo`. When a cached response is returned, if we attempt to use the info right away to start a download, Safari will skip it. There is no apparent error logged, so it's unclear why.

If we wait until the next tick to return the cached results, Safari will download as expected. (Even returning the cached response with `Promise.resolve` doesn't work.)

Maybe this is related to Safari's security around user gestures for file downloads. Or maybe not. It's so strange that I'm not even convinced this is the root issue.