rollback vite to v4, and instead use deprecated `assert` in @umn-cla/upp-aws-s3
`@umn-cla/upp-aws-s3` uses `import ... with` syntax which causes `vite dev` to fail with error since vite 4 doesn't support it yet.

Originally, I bumped to latest vite (v7), but this had a few too many changes (e.g. manifest.json location) and broke the build. So, it'd be better to do this separate from the UI admin pages.

Instead, the `@umn-cla/uppy-aws-s3` package was modified to use `import ... assert` instead of `import ... with` which is supported.

This then:
- reverts vite to v4
- changes to "type: module" (which fixes another build issue). And updates config using the older commonjs import syntax.
- bumps the uppy-aws package to the version with `assert`