fix flash when saving and navigation from editAsset to createAsset
previously the app would flash when a save happened because there was `:key="$route.fullPath"` on `<routerview>`  causing a full rerender when routes changed.

This removed ths full rerender, which solved the flash, but then re-introduced the issue when navigating from Edit Asset page to Create Asset page - no state would be reset.

This is because the components used the same component, but different routes in the router. This consolidates routes, which then allows us to use an onBeforeRouteUpdate hook to reset data.

For good measure, we throw in a confirmation dialog to make sure the user knows they have unsaved data and are about to create a new asset.