debounce save asset to avoid race condition
With multiple asset uploads, we could have a race condition where an older saveAsset request completes AFTER a newer one, resulting in us losing the newer information.

To mitigate this, we debounce saveAsset – so that it's called at most every 500ms.

`saveAsset` references localAsset.value directly, so it will use the most recent data when it finally executes.