fix: normalize the multiselect fieldContent keys
This addresses an issue where multiselect widget fieldContent keys were
being saved with "spaces". The legacy data does not have spaces in keys.

We now normalize the keys to be alphanumeric with no spaces.

To properly test this, we extracted the `useCascadeSelect` logic into
its own composable, and then create a separate `SimpleCascadeSelect`
component using that composable. The existing `CascadeSelect` component,
which is used by advanced search, is left unchange for now.

Internally, the new `useCascadeSelect` flattens the complex nested
cascade options object for easier searching and filtering.