update modelValue on every input rather than commit
Previously, we were committing changes to modelValue once a confirmed selection is made with enter, tab, or `,`.

However, this causes a ux issue with the textWidget: if the user edits text in an autcomplete widget, they need to click away or enter or something before they're able to save.

This changes the Autocomplete so that `modelValue` always represents the current input. And, if a selection is made, it emits a modelValue update.

Then, the responsibility for committing the change on `enter` or `,` or `tab` is moved to the EditTagWidget.